remove : DrawableBitmap.ts
createjsのサポートを終了する。
