+++ bin/DrawingMode.d.ts
@@ -1,0 +1,5 @@
+export declare enum DrawingMode {
+    pen = "pen",
+    eraser = "eraser"
+}
+//# sourceMappingURL=DrawingMode.d.ts.map
