+++ bin/DrawingMode.js
@@ -1,0 +1,5 @@
+export var DrawingMode;
+(function (DrawingMode) {
+    DrawingMode["pen"] = "pen";
+    DrawingMode["eraser"] = "eraser";
+})(DrawingMode || (DrawingMode = {}));
