+++ bin/DrawingOption.d.ts
@@ -1,0 +1,7 @@
+import { DrawingMode } from './DrawingMode';
+export declare class DrawingOption {
+    mode: DrawingMode;
+    color?: string;
+    width?: number;
+}
+//# sourceMappingURL=DrawingOption.d.ts.map
