+++ bin/DrawingOption.js
@@ -1,0 +1,2 @@
+export class DrawingOption {
+}
