+++ bin/index.d.ts
@@ -1,3 +1,4 @@
-export { DrawableBitmap, DrawingMode, DrawingOption } from "./DrawableBitmap";
+export * from "./DrawingMode";
+export * from "./DrawingOption";
 export * from "./DrawableCanvas";
 //# sourceMappingURL=index.d.ts.map
