+++ bin/index.js
@@ -1,2 +1,3 @@
-export { DrawableBitmap, DrawingMode, DrawingOption } from "./DrawableBitmap";
+export * from "./DrawingMode";
+export * from "./DrawingOption";
 export * from "./DrawableCanvas";
