+++ package.json
@@ -8,12 +8,10 @@
   "license": "MIT",
   "devDependencies": {
     "@masatomakino/gulptask-demo-page": "^0.6.7",
-    "@types/easeljs": "^1.0.1",
     "browser-sync": "^2.29.3",
     "husky": "^8.0.3",
     "prettier": "^2.8.8",
     "pretty-quick": "^3.1.3",
-    "ts-node": "^10.9.1",
     "typedoc": "^0.24.8",
     "typescript": "^5.1.6"
   },
@@ -22,7 +20,7 @@
     "build": "npm run buildTS && npm run demo && npm run typedocs",
     "buildTS": "tsc --project tsconfig.json",
     "watch:tsc": "tsc -W --incremental --project tsconfig.json",
-    "demo": "npx @masatomakino/gulptask-demo-page --externalScripts 'https://code.createjs.com/1.0.0/easeljs.min.js' ",
+    "demo": "npx @masatomakino/gulptask-demo-page",
     "watch:demo": "npm run demo -- -W",
     "typedocs": "npx typedoc --out ./docs/api src/index.ts",
     "server": "npx browser-sync ./docs/demo -w",
