+++ src/DrawingMode.ts
@@ -1,0 +1,4 @@
+export enum DrawingMode {
+  pen = "pen",
+  eraser = "eraser",
+}
