+++ src/DrawingOption.ts
@@ -1,0 +1,7 @@
+import { DrawingMode } from "./DrawingMode";
+
+export class DrawingOption {
+  mode: DrawingMode;
+  color?: string;
+  width?: number;
+}
